// generated/src/app/api-models/special-contract-header.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SpecialContractHeader {
  seqSpecCont: number;
  priceRule: string;
  seqProvId: number;
  lineOfBusiness: string;
  panelId: string;
  effectiveDate: Date;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}